requirements.txt
pyflakes
djangorestframework